+++ core/src/main/kotlin/org/agrfesta/sh/api/core/application/ports/outbounds/devices/OnOffActuator.kt
@@ -2,10 +2,9 @@ package org.agrfesta.sh.api.core.application.ports.outbounds.devices
 
 import arrow.core.Either
 import org.agrfesta.sh.api.core.domain.devices.ActuatorStatus
+import org.agrfesta.sh.api.core.domain.failures.ActuatorOperationFailure
 
 interface Actuator : DeviceDriver
-
-interface ActuatorOperationFailure
 
 interface OnOffActuator : Actuator {
     fun getActuatorStatus(): Either<ActuatorOperationFailure, ActuatorStatus>